Based on your comments I've updated the notes (on the erm) and updated the following things about the erm: - Changed the naming of the Feedback* relations so they are not as confusing - Updated SellerFeedbackForBuyer to remove the rating. - Added in ItemListing "buyer", which will be a username. Even though it references a userame, it can't be a foreign key - i believe - because the value will initially be empty since there is not a buyer yet. - The reason for adding "buyer": --- Make the following type of query faster: Display all feedback for the current user that any seller has left for them for any items that they have won the bidding to (i think that is the correct statement).
This brings up another point - mainly just so I don't forget about
this. We have have at least one case of things happening behind the
scenes: The buyer of an item after the auction has closed. This will
need to be updated in the database when the current time of the site
is beyond the end time of the auction. The point is - I will get
around to adding a common php function that should be called when any
activity is done (since this is how we update the site time) to update
any item listings that need to have the buyer field updated. So for
example (just throwing my thoughts out there):

we could have a common function in dbinfo.php such that you could do
the following:

$dbinfo->save_useractivity($activity_description);

and in dbinfo_t save_useractivity($activity_description) would be
defined somethign like:

{
$user = $this->user();
$day = $this->day();
$hour...
$minute..
$this->query("insert into ... activity=$activity_description, username=$user, day=$day, hour....");
$this->increment_site_time();
$this->update_closed_item_listings();
etc...
}

Since anything that has the potential to add information to the
database (bidding, editing profile, listing an item, etc..) will be
saved as a user activity as well -> which will advance the time, I
think this is a good catch-all place to put something like this.

I'll think about the specifics of it and worry about it when we start
our php stuff.